+++ CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.4.2] - 2021-08-21
+## [3.4.3] - 2021-08-23
 ### Fixed
 - 3.4.1: Attempt to fix the issues with Patreon login/connect
 - 3.4.2: Patreon UI for Warhammer system
+- 3.4.3: Moulinette Picker still enabled when configuration is off
 ### Added
 - Support for Moulinette Cloud (all)
 - Progressbar for Cloud update